agent/pyproject.toml
[project]
name = "agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "bedrock-agentcore>=0.1.3",
    "bedrock-agentcore-starter-toolkit>=0.1.10",
    "boto3>=1.40.12",
    "strands-agents>=1.5.0",
    "strands-agents-tools>=0.2.7",
    "uv>=0.8.17",
]
